simplify external settings api
Move setting type constants and binding contracts onto ExternalSettingEntry.

Inline the external settings slot factory so the API has fewer one-use classes while keeping the same page registration behavior.
