--- src/main/java/journeymap/client/api/settings/BooleanSettingBinding.java
@@ -1,10 +1,0 @@
-package journeymap.client.api.settings;
-
-public interface BooleanSettingBinding
-{
-    boolean get();
-
-    void set(boolean value);
-
-    boolean getDefaultValue();
-}
