--- src/main/java/journeymap/client/api/settings/DoubleSettingBinding.java
@@ -1,20 +1,0 @@
-package journeymap.client.api.settings;
-
-public interface DoubleSettingBinding
-{
-    double get();
-
-    void set(double value);
-
-    double getDefaultValue();
-
-    double getMinValue();
-
-    double getMaxValue();
-
-    double getStep();
-
-    int getPrecision();
-
-    String getSuffix();
-}
