--- src/main/java/journeymap/client/api/settings/ExternalSettingKind.java
@@ -1,12 +1,0 @@
-package journeymap.client.api.settings;
-
-public enum ExternalSettingKind
-{
-    LABEL,
-    BOOLEAN,
-    INT_SLIDER,
-    DOUBLE_SLIDER,
-    ENUM_LIST,
-    STRING_LIST,
-    STRING_INPUT
-}
