--- src/main/java/journeymap/client/api/settings/IntSettingBinding.java
@@ -1,18 +1,0 @@
-package journeymap.client.api.settings;
-
-public interface IntSettingBinding
-{
-    int get();
-
-    void set(int value);
-
-    int getDefaultValue();
-
-    int getMinValue();
-
-    int getMaxValue();
-
-    int getStep();
-
-    String getSuffix();
-}
