--- src/main/java/journeymap/client/api/settings/ListSettingBinding.java
@@ -1,16 +1,0 @@
-package journeymap.client.api.settings;
-
-import java.util.List;
-
-public interface ListSettingBinding<T>
-{
-    T get();
-
-    void set(T value);
-
-    T getDefaultValue();
-
-    List<T> getOptions();
-
-    String getLabel(T value);
-}
