--- src/main/java/journeymap/client/api/settings/SettingsPageProvider.java
@@ -1,9 +1,0 @@
-package journeymap.client.api.settings;
-
-/**
- * Supplies a settings page that JourneyMap can display inside its options GUI.
- */
-public interface SettingsPageProvider
-{
-    SettingsPageDefinition getPageDefinition();
-}
