--- src/main/java/journeymap/client/api/settings/StringInputBinding.java
@@ -1,12 +1,0 @@
-package journeymap.client.api.settings;
-
-public interface StringInputBinding
-{
-    String get();
-
-    void set(String value);
-
-    String getDefaultValue();
-
-    int getMaxLength();
-}
